requirements.txt
# prompt: make full req text file with minor details 

streamlit
scikit-learn
pandas
numpy
joblib
matplotlib 
seaborn 
sklearn.ensemble
sklearn.metrics